renamed components for genre filtering to be more accurate
