+++ client/src/components/GenreFilterList.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import FilterListItem from './FilterListItem';
+import GenreFilterListItem from './GenreFilterListItem';
 import { makeStyles } from '@material-ui/core/styles';
 
 const useStyles = makeStyles(theme => ({
@@ -15,7 +15,7 @@ export default function Filter(props) {
   // map over keys of songsByGenres object to render genre chips
   const list = Object.keys(props.songsByGenre).map((genre, index) => {
     return (
-        <FilterListItem
+        <GenreFilterListItem
           key={index}
           genreName={genre}
           selected={props.value.includes(genre) ? true : false}
@@ -28,7 +28,7 @@ export default function Filter(props) {
     <div className={classes.root}>
       <ul>
         {list}
-        <FilterListItem genreName='All'/>
+        <GenreFilterListItem genreName='All'/>
       </ul>
     </div>
   );
